+++ src/components/molecules/Button/utils/getButtonStyles/utils/constants/index.ts
@@ -1,6 +1,6 @@
 const verticalHeights = ['top', 'bottom'];
 const validTypes = ['main', 'secondary'];
-const validVariants = ['contained', 'outlined', 'text'];
+const validVariants = ['contained', 'outlined', 'text', 'white'];
 const validIconPositions = ['top', 'bottom', 'left', 'right'];
 
 export {verticalHeights, validTypes, validVariants, validIconPositions};